+++ source/main.py
@@ -3,6 +3,7 @@
 
 # Import libraries
 import streamlit as st
+import hmac
 import pandas as pd
 import plotly.graph_objects as go
 import scraper
@@ -11,6 +12,31 @@ import ASCE
 import plotting
 import make_dataframe
 
+def check_password():
+    """Returns `True` if the user had the correct password."""
+
+    def password_entered():
+        """Checks whether a password entered by the user is correct."""
+        if hmac.compare_digest(st.session_state["password"], st.secrets["password"]):
+            st.session_state["password_correct"] = True
+            del st.session_state["password"]  # Don't store the password.
+        else:
+            st.session_state["password_correct"] = False
+
+    # Return True if the passward is validated.
+    if st.session_state.get("password_correct", False):
+        return True
+
+    # Show input for password.
+    st.text_input(
+        "Password", type="password", on_change=password_entered, key="password"
+    )
+    if "password_correct" in st.session_state:
+        st.error("Password incorrect")
+    return False
+
+if not check_password():
+    st.stop()  # Do not continue if check_password is not True.
 # Define a function to show plot, dataframe, and URLs
 def do_all_the_processing():
     # Get the value of the slider